Reorganizing some code -- thinking ahead to a more DRY approach
