+++ src/Looper.cpp
@@ -2,6 +2,7 @@
 #include "osdialog.h"
 #include "Common/sample.hpp"
 #include "Common/VoxglitchSamplerModule.hpp"
+#include "Common/VoxglitchSamplerModuleWidget.hpp"
 #include "settings.hpp"
 
 #include "Looper/defines.h"
