+++ src/Looper/LooperWidget.hpp
@@ -1,4 +1,4 @@
-struct LooperWidget : ModuleWidget
+struct LooperWidget : VoxglitchSamplerModuleWidget
 {
   LooperWidget(Looper* module)
   {
@@ -18,40 +18,6 @@ struct LooperWidget : ModuleWidget
     looper_waveform_display->module = module;
     addChild(looper_waveform_display);
   }
-
-  struct InterpolationOffOption : MenuItem {
-    Looper *module;
-
-    void onAction(const event::Action &e) override {
-      module->interpolation = 0;
-    }
-  };
-
-  struct InterpolationLinearOption : MenuItem {
-    Looper *module;
-
-    void onAction(const event::Action &e) override {
-      module->interpolation = 1;
-    }
-  };
-
-  struct SampleInterpolationMenuItem : MenuItem {
-    Looper *module;
-
-    Menu *createChildMenu() override {
-      Menu *menu = new Menu;
-
-      InterpolationOffOption *interpolation_off_option = createMenuItem<InterpolationOffOption>("Off", CHECKMARK(module->interpolation == 0));
-      interpolation_off_option->module = module;
-      menu->addChild(interpolation_off_option);
-
-      InterpolationLinearOption *interpolation_linear_option = createMenuItem<InterpolationLinearOption>("Linear", CHECKMARK(module->interpolation == 1));
-      interpolation_linear_option->module = module;
-      menu->addChild(interpolation_linear_option);
-
-      return menu;
-    }
-  };
 
   void appendContextMenu(Menu *menu) override
   {
